use real tts trainer
